+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.0.4
 * [IMP] Annotate `GYWBtDevice.fbDevice` and `GYWBtDevice.findCharacteristic` with `@internal`
 * [FIX] Fix `GYWBtDevice.connect` returning true when connection fails
+* [FIX] Fix `GYWStatusException: The device is already trying to be disconnected.`
 
 ## 2.0.3
 * [FIX] Fix the use of images that are not library icons